requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0
fastapi>=0.104.1
uvicorn>=0.24.0
httpx>=0.25.1
python-multipart>=0.0.6
